[feat] #60 지도 관련된 공통 함수 utility 모듈로 이동
