+++ core/utility/build.gradle
@@ -36,6 +36,12 @@ android {
 }
 
 dependencies {
+    implementation project(":core:designsystem")
+    implementation project(":domain")
+
+    implementation libs.kakao.map
+    implementation libs.play.services.maps
+    implementation libs.play.services.location
     implementation libs.timber
     implementation libs.androidx.core.ktx
     implementation libs.androidx.activity.compose
