+++ core/utility/src/main/java/com/yapp/buddycon/utility/Permission.kt
@@ -1,0 +1,16 @@
+package com.yapp.buddycon.utility
+
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import androidx.core.content.ContextCompat
+
+fun checkLocationPermission(
+    context: Context
+): Boolean {
+    val permissions = arrayOf(
+        Manifest.permission.ACCESS_COARSE_LOCATION,
+        Manifest.permission.ACCESS_FINE_LOCATION
+    )
+    return permissions.all { ContextCompat.checkSelfPermission(context, it) == PackageManager.PERMISSION_GRANTED }
+}
