+++ feature/gifticon/build.gradle
@@ -40,6 +40,7 @@ android {
 dependencies {
     implementation project(":core:designsystem")
     implementation project(":domain")
+    implementation project(":core:utility")
 
     implementation libs.hilt.android
     kapt libs.hilt.compiler
